requirements.txt
azure-functions
grpcio
grpcio-tools
newspaper3k
pymongo
dnspython